+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - General:
   - Support for time stamp in the date (i.e. 2023-04-02T09:22-03:00)
   - Support to pass variables to the 3D models download URL (#414)
+  - Support for netclass flags (#418)
 - Command line:
   - `--banner N` Option to display a banner
   - `--log FILE` Option to log to a file, in addition to the stderr